Move `DrawingOrErasingView` content to `App`
