+++ frontend/src/components/App/index.tsx
@@ -1,4 +1,5 @@
-import React from 'react';
+import React, { useMemo } from 'react';
+import classnames from 'classnames';
 import {
   Mode,
   fillCell,
@@ -6,17 +7,67 @@ import {
   switchToErasing,
   switchToDrawing,
   editCrosswordAndAbortFetch,
+  switchToEnteringQuestions,
+  generateQuestions,
 } from 'store';
 import useAppDispatch from 'hooks/useAppDispatch';
 import useAppSelector from 'hooks/useAppSelector';
-import DrawingOrErasingView from 'views/DrawingOrErasingView';
+import { Mode as GridMode } from 'components/Grid';
+import Button from 'components/Button';
+import GridWrapper from 'components/GridWrapper';
+import Label, { LabelSize } from 'components/Label';
 import Layout from 'components/Layout';
+import Tabs from 'components/Tabs';
+import Square from 'icons/Square';
+import './styles.scss';
 
 function App() {
   const dispatch = useAppDispatch();
   const { grid, mode, fetchAbortController, apiFailed } = useAppSelector(
     (state) => state.general,
   );
+
+  const getTabByModeAndIsSelected = (
+    currentMode: Mode.Draw | Mode.Erase,
+    isSelected: boolean,
+  ) => {
+    const getLabel = (labelMode: Mode.Draw | Mode.Erase) => {
+      const labelByModeAndIsSelected = {
+        [Mode.Draw]: {
+          regular: 'Draw',
+          selected: 'Drawing',
+        },
+        [Mode.Erase]: {
+          regular: 'Erase',
+          selected: 'Erasing',
+        },
+      };
+
+      return labelByModeAndIsSelected[labelMode][
+        isSelected ? 'selected' : 'regular'
+      ];
+    };
+
+    const label = getLabel(currentMode);
+    const alternativeLabel = getLabel(
+      currentMode === Mode.Draw ? Mode.Erase : Mode.Draw,
+    );
+
+    return {
+      label,
+      alternativeLabel,
+      icon: <Square isFilled={currentMode === Mode.Erase} />,
+    };
+  };
+
+  const booleanGrid = useMemo(
+    () => grid.map((row) => row.map((cell) => !!cell)),
+    [grid],
+  );
+
+  const isGridEmpty = useMemo(() => {
+    return booleanGrid.every((row) => row.every((cell) => !cell));
+  }, [booleanGrid]);
 
   const getLoaderLabel = (): string | null => {
     if (fetchAbortController) {
@@ -32,42 +83,83 @@ function App() {
     return null;
   };
 
-  const getCurrentView = () => {
-    switch (mode) {
-      case Mode.Draw:
-      case Mode.Erase:
-        return (
-          <DrawingOrErasingView
-            mode={mode}
-            grid={grid}
-            onModeChange={() =>
-              dispatch(
-                (mode === Mode.Draw ? switchToErasing : switchToDrawing)(),
-              )
-            }
-            onCellChange={(row, column) =>
-              dispatch(
-                (mode === Mode.Draw ? fillCell : eraseCell)({ row, column }),
-              )
-            }
-            loaderLabel={getLoaderLabel()}
-            onEditTabClick={
-              fetchAbortController &&
-              (() => dispatch(editCrosswordAndAbortFetch()))
-            }
-          />
-        );
-      case Mode.EnterQuestions:
-        return <></>;
-      case Mode.Answer:
-        return <></>;
-      case Mode.Puzzle:
-        return <></>;
-    }
-  };
-
   return (
-    <Layout title='Crossword Generator & Solver'>{getCurrentView()}</Layout>
+    <Layout title='Crossword Generator & Solver'>
+      <GridWrapper
+        gridProps={{
+          matrix: booleanGrid,
+          mode: mode === Mode.Draw ? GridMode.Draw : GridMode.Erase,
+          onChange: (row, column) =>
+            dispatch(
+              (mode === Mode.Draw ? fillCell : eraseCell)({
+                row,
+                column,
+              }),
+            ),
+        }}
+        loaderLabel={getLoaderLabel()}
+      >
+        <Tabs
+          onEditClick={
+            fetchAbortController
+              ? () => dispatch(editCrosswordAndAbortFetch())
+              : undefined
+          }
+          selectedTab={
+            mode === Mode.Draw || mode === Mode.Erase
+              ? getTabByModeAndIsSelected(mode, true)
+              : undefined
+          }
+          secondaryTab={
+            mode === Mode.Draw || mode === Mode.Erase
+              ? {
+                  ...getTabByModeAndIsSelected(
+                    mode === Mode.Draw ? Mode.Erase : Mode.Draw,
+                    false,
+                  ),
+                  onClick: () =>
+                    dispatch(
+                      (mode === Mode.Draw
+                        ? switchToErasing
+                        : switchToDrawing)(),
+                    ),
+                }
+              : undefined
+          }
+        />
+      </GridWrapper>
+      <div className='option-buttons-wrapper'>
+        <div
+          className={classnames(
+            'option-buttons-wrapper-item',
+            (fetchAbortController || !isGridEmpty) &&
+              'option-buttons-wrapper-item_hidden',
+          )}
+        >
+          <Label
+            content='Let’s draw some squares first!'
+            size={LabelSize.Large}
+          />
+        </div>
+        <div
+          className={classnames(
+            'option-buttons-wrapper-item',
+            (fetchAbortController || isGridEmpty) &&
+              'option-buttons-wrapper-item_hidden',
+            'option-buttons',
+          )}
+        >
+          <Button
+            label='Generate questions'
+            onClick={() => dispatch(generateQuestions())}
+          />
+          <Button
+            label='Enter questions & solve'
+            onClick={() => dispatch(switchToEnteringQuestions())}
+          />
+        </div>
+      </div>
+    </Layout>
   );
 }
