--- frontend/src/views/DrawingOrErasingView/styles.scss
@@ -1,43 +1,0 @@
-.drawing-or-erasing-view {
-    display: contents;
-
-    &__buttons-wrapper {
-        display: grid;
-        grid-template-areas: 'buttonWrapperItem';
-        place-items: center;
-        width: 100%;
-        height: 100%;
-
-        &-item {
-            grid-area: buttonWrapperItem;
-            transition: all var(--transition-duration) var(--transition-function);
-
-            &_hidden {
-                opacity: 0;
-                visibility: hidden;
-            }
-
-            &:not(&_hidden) {
-                opacity: 1;
-                visibility: visible;
-            }
-        }
-    }
-
-    &__buttons {
-        display: flex;
-        gap: 27px;
-        flex-direction: column;
-        width: 100%;
-        padding: 0 48px;
-        box-sizing: border-box;
-
-        & > *:first-child {
-            align-self: flex-start;
-        }
-
-        & > *:last-child {
-            align-self: flex-end;
-        }
-    }
-}
